+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Add unindenting of ``{% extends %}`` tags, and top-level ``{% block %}`` and ``{% endblock %}`` tags where ``{% extends %}`` is used.
+
+  `PR #30 <https://github.com/adamchainz/djade/pull/30>`__.
+
 0.1.0 (2024-09-21)
 ------------------